2018.11.9 v0.0.1, 优化文档 #16
