+++ docs/app.rst
@@ -1,0 +1,10 @@
+``app``
+=============================
+
+.. autosummary::
+    app.Summer
+    app.Summer.load_conf
+    app.Summer.create_server
+
+.. automodule:: app
+    :members:
