+++ docs/change_log.rst
@@ -1,2 +1,5 @@
 更新记录
 ===========================
+2018.11.09 v0.0.1
+---------------------------
+* 首次发布;
