--- docs/command.rst
@@ -1,13 +1,0 @@
-``command``
-=============================
-
-.. autosummary::
-    command.cli
-    command.start_server
-    command.create
-    command._load_config
-    command._create_folder
-    command._create_file
-
-.. automodule:: command
-    :members:
