+++ summer/tpl/__init__.tpl
@@ -1,3 +1,4 @@
+import os
 from summer import *
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
