+++ summer/tpl/requirements.tpl
@@ -1,3 +1,3 @@
 SQLAlchemy==1.2.1
 pytest==3.6.2
-py_summer==0.0.1
+summer==0.0.1
